# api/openapi.yaml
openapi: 3.0.3
info:
  title: Switchyard Local API
  version: 1.0.0
  description: Local control-plane contract shared by browser, desktop, and CLI clients.
servers:
  - url: /api/v1
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      summary: List durable multi-project workspaces
      responses:
        "200":
          description: Workspaces in stable display order
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Workspace" }
        default:
          $ref: "#/components/responses/Problem"
    post:
      operationId: createWorkspace
      summary: Create a validated workspace dependency graph
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceDefinition" }
      responses:
        "201":
          description: Workspace created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        default:
          $ref: "#/components/responses/Problem"
  /workspaces/{workspaceId}:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    get:
      operationId: getWorkspace
      summary: Read one workspace graph and its latest execution
      responses:
        "200":
          description: Workspace graph and execution state
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        default:
          $ref: "#/components/responses/Problem"
    put:
      operationId: updateWorkspace
      summary: Replace a workspace graph after revision validation
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceUpdate" }
      responses:
        "200":
          description: Updated workspace
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workspace" }
        default:
          $ref: "#/components/responses/Problem"
    delete:
      operationId: deleteWorkspace
      summary: Remove workspace metadata without touching projects or runtime data
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      responses:
        "204": { description: Workspace metadata removed }
        default:
          $ref: "#/components/responses/Problem"
  /workspaces/{workspaceId}/operations:
    parameters:
      - $ref: "#/components/parameters/WorkspaceId"
    post:
      operationId: createWorkspaceOperation
      summary: Queue an ordered workspace start or dependency-safe bulk stop
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkspaceOperationRequest" }
      responses:
        "202":
          description: Durable workspace operation accepted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Operation" }
        default:
          $ref: "#/components/responses/Problem"
components:
  parameters:
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
  responses:
    Problem:
      description: RFC 9457-style problem details
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
  schemas:
...
    WorkspaceFailurePolicy:
      type: string
      enum: [rollback, continue]
...
    WorkspaceMemberRole:
      type: string
      enum: [application, dependency, tooling]
    WorkspaceProjectStatus:
      type: string
      enum: [idle, queued, blocked, starting, checking_health, running, start_failed, stopping, stopped, stop_failed, rolling_back, rolled_back, rollback_failed, cancelled]
    WorkspaceMemberDefinition:
      type: object
      additionalProperties: false
      required: [projectId, role, order, healthGate, healthTimeoutSeconds]
      properties:
        projectId: { type: string, minLength: 1, maxLength: 128 }
        role: { $ref: "#/components/schemas/WorkspaceMemberRole" }
        order: { type: integer, minimum: 0, maximum: 10000 }
        healthGate: { type: boolean }
        healthTimeoutSeconds: { type: integer, minimum: 0, maximum: 3600 }
    WorkspaceMember:
      type: object
      additionalProperties: false
      required: [projectId, role, order, healthGate, healthTimeoutSeconds, status]
      properties:
        projectId: { type: string, minLength: 1, maxLength: 128 }
        role: { $ref: "#/components/schemas/WorkspaceMemberRole" }
        order: { type: integer, minimum: 0, maximum: 10000 }
        healthGate: { type: boolean }
        healthTimeoutSeconds: { type: integer, minimum: 0, maximum: 3600 }
        status: { $ref: "#/components/schemas/WorkspaceProjectStatus" }
        message: { type: string, maxLength: 2000 }
    WorkspaceDependency:
      type: object
      additionalProperties: false
      required: [projectId, dependsOnProjectId]
      properties:
        projectId: { type: string, minLength: 1, maxLength: 128 }
        dependsOnProjectId: { type: string, minLength: 1, maxLength: 128 }
    WorkspaceRecipe:
      type: object
      additionalProperties: false
      required: [id, name, kind, order, arguments]
      properties:
        id: { type: string, minLength: 1, maxLength: 128 }
        name: { type: string, minLength: 1, maxLength: 160 }
        kind: { type: string, enum: [open_url, open_terminal, open_editor, start_agent] }
        projectId: { type: string, minLength: 1, maxLength: 128 }
        target: { type: string, maxLength: 2048 }
        arguments: { type: array, maxItems: 32, items: { type: string, maxLength: 2048 } }
        order: { type: integer, minimum: 0, maximum: 10000 }
    WorkspaceProfile:
      type: object
      additionalProperties: false
      required: [id, name, projectIds, maxParallel, lowMemory]
      properties:
        id: { type: string, minLength: 1, maxLength: 128 }
        name: { type: string, minLength: 1, maxLength: 160 }
        description: { type: string, maxLength: 1000 }
        projectIds: { type: array, maxItems: 128, items: { type: string, minLength: 1, maxLength: 128 } }
        maxParallel: { type: integer, minimum: 1, maximum: 64 }
        lowMemory: { type: boolean }
        memoryBudgetBytes: { type: integer, format: int64, minimum: 0 }
    WorkspaceDefinition:
      type: object
      additionalProperties: false
      required: [name, policy, members, dependencies, recipes, profiles]
      properties:
        name: { type: string, minLength: 1, maxLength: 160 }
        description: { type: string, maxLength: 2000 }
        policy: { $ref: "#/components/schemas/WorkspaceFailurePolicy" }
        profile: { type: string, maxLength: 128 }
        members: { type: array, minItems: 1, maxItems: 128, items: { $ref: "#/components/schemas/WorkspaceMemberDefinition" } }
        dependencies: { type: array, maxItems: 512, items: { $ref: "#/components/schemas/WorkspaceDependency" } }
        recipes: { type: array, maxItems: 128, items: { $ref: "#/components/schemas/WorkspaceRecipe" } }
        profiles: { type: array, maxItems: 64, items: { $ref: "#/components/schemas/WorkspaceProfile" } }
    WorkspaceUpdate:
      type: object
      additionalProperties: false
      required: [name, policy, members, dependencies, recipes, profiles, revision]
      properties:
        name: { type: string, minLength: 1, maxLength: 160 }
        description: { type: string, maxLength: 2000 }
        policy: { $ref: "#/components/schemas/WorkspaceFailurePolicy" }
        profile: { type: string, maxLength: 128 }
        members: { type: array, minItems: 1, maxItems: 128, items: { $ref: "#/components/schemas/WorkspaceMemberDefinition" } }
        dependencies: { type: array, maxItems: 512, items: { $ref: "#/components/schemas/WorkspaceDependency" } }
        recipes: { type: array, maxItems: 128, items: { $ref: "#/components/schemas/WorkspaceRecipe" } }
        profiles: { type: array, maxItems: 64, items: { $ref: "#/components/schemas/WorkspaceProfile" } }
        revision: { type: integer, format: int64, minimum: 1 }
    WorkspaceProjectResult:
      type: object
      additionalProperties: false
      required: [projectId, role, status, order]
      properties:
        projectId: { type: string }
        role: { $ref: "#/components/schemas/WorkspaceMemberRole" }
        status: { $ref: "#/components/schemas/WorkspaceProjectStatus" }
        message: { type: string }
        order: { type: integer }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
    WorkspaceExecution:
      type: object
      additionalProperties: false
      required: [id, workspaceId, kind, state, policy, removeData, projects, startedAt]
      properties:
        id: { type: string }
        workspaceId: { type: string }
        kind: { type: string, enum: [start, stop] }
        state: { type: string, enum: [running, succeeded, partially_succeeded, failed, cancelled] }
        policy: { $ref: "#/components/schemas/WorkspaceFailurePolicy" }
        profileId: { type: string }
        removeData: { type: boolean }
        projects: { type: array, items: { $ref: "#/components/schemas/WorkspaceProjectResult" } }
        errorMessage: { type: string }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: string, format: date-time }
    Workspace:
      type: object
      additionalProperties: false
      required: [id, name, policy, members, dependencies, recipes, profiles, revision, createdAt, updatedAt]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        policy: { $ref: "#/components/schemas/WorkspaceFailurePolicy" }
        profile: { type: string }
        members: { type: array, items: { $ref: "#/components/schemas/WorkspaceMember" } }
        dependencies: { type: array, items: { $ref: "#/components/schemas/WorkspaceDependency" } }
        recipes: { type: array, items: { $ref: "#/components/schemas/WorkspaceRecipe" } }
        profiles: { type: array, items: { $ref: "#/components/schemas/WorkspaceProfile" } }
        lastRun: { $ref: "#/components/schemas/WorkspaceExecution" }
        revision: { type: integer, format: int64, minimum: 1 }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    WorkspaceOperationRequest:
      type: object
      additionalProperties: false
      required: [action]
      properties:
        action: { $ref: "#/components/schemas/WorkspaceOperationAction" }
        policy: { $ref: "#/components/schemas/WorkspaceFailurePolicy" }
        profileId: { type: string, maxLength: 128 }
        removeData: { type: boolean, default: false }
        confirmDataRemoval: { type: boolean, default: false }
        runRecipes: { type: boolean, default: false }
    WorkspaceOperationAction:
      type: string
      enum: [start, stop]
      x-enum-varnames: [WorkspaceStart, WorkspaceStop]
    Operation:
      type: object
      additionalProperties: false
      required:
        - id
        - projectId
        - kind
        - state
        - requestedAt
        - updatedAt
        - cancellationRequested
      properties:
        id:
          type: string
        projectId:
          type: string
        workspaceId:
          type: string
        kind:
          type: string
        state:
          $ref: "#/components/schemas/OperationState"
        errorCode:
          type: string
        errorMessage:
          type: string
        cancellationRequested:
          type: boolean
        requestedAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
        finishedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    OperationState:
      type: string
      enum: [queued, running, succeeded, failed, cancelled, partially_succeeded]
...
    ProblemDetails:
      type: object
      additionalProperties: false
      required: [type, title, status, code, correlationId]
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
          minimum: 400
          maximum: 599
        detail:
          type: string
        code:
          type: string
        correlationId:
          type: string